# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API utilizada no site tickets.zauso-dev.com.br",
        "title": "API Banners",
        "contact": {},
        "version": "1.0"
    },
    "host": "api-banners.zauso-dev.com.br",
    "basePath": "/",
    "paths": {
        "/banners": {
            "get": {
                "description": "Retorna banners cadastrados",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Banners"
                ],
                "summary": "Lista banners",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Banners"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Cria novo banner",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Banners"
                ],
                "summary": "Cria Banner",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Evento",
                        "name": "id_evento",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Titulo do Banner",
                        "name": "titulo",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Status do banner",
                        "name": "ativo",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Imagem do banner",
                        "name": "imagem",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Banners"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/banners/{id}": {
            "delete": {
                "description": "Remove banners pelo ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Banners"
                ],
                "summary": "Remove banners",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Banner",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MessageResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Banners": {
            "type": "object",
            "properties": {
                "ativo": {
                    "type": "boolean"
                },
                "data_criacao": {
                    "type": "string"
                },
                "id_banner": {
                    "type": "integer"
                },
                "id_evento": {
                    "type": "integer"
                },
                "imagem": {
                    "type": "string"
                },
                "titulo": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}